refactor(ms-word-addin): create adapt toolbar widget
