--- apps/ms-word-addin/src/features/readingTools/index.ts
@@ -1,1 +1,0 @@
-export { default as ReadingTools } from './ui/ReadingTools.vue'
