+++ apps/ms-word-addin/src/features/readingToolsToolbar/index.ts
@@ -1,0 +1,1 @@
+export { default as ReadingToolsToolbar } from './ui/ReadingToolsToolbar.vue'
