+++ apps/ms-word-addin/src/features/readingToolsToolbar/model/useMousePosition.ts
