+++ apps/ms-word-addin/src/features/readingToolsToolbar/ui/ReadingToolsMask.vue
