+++ apps/ms-word-addin/src/features/readingToolsToolbar/ui/ReadingToolsRuler.vue
