+++ apps/ms-word-addin/src/widgets/adaptToolbar/index.ts
@@ -1,0 +1,1 @@
+export { default as AdaptToolbar } from './ui/AdaptToolbar.vue'
