+++ apps/ms-word-addin/src/widgets/adaptToolbar/ui/AdaptToolbar.vue
@@ -1,0 +1,9 @@
+<script setup lang="ts">
+import { ReadingToolsToolbar } from '@/features/readingToolsToolbar'
+import PrintButton from '@/shared/ui/PrintButton.vue'
+</script>
+<template>
+  <ReadingToolsToolbar>
+    <PrintButton />
+  </ReadingToolsToolbar>
+</template>
